% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavi_fa.R
\name{cavi_fa}
\alias{cavi_fa}
\title{Coordinate Ascent Variational Inference for Bayesian Factor Analysis}
\usage{
cavi_fa(
  X,
  J,
  verbose = 1,
  tol = 0.01,
  min_iter = 5,
  max_iter = 100,
  scale = TRUE
)
}
\arguments{
\item{X}{A data matrix, of dimensions \eqn{P \times N}}

\item{J}{Number of factors.}

\item{verbose}{Desired verbosity of output printed out each iteration. A value greater than 1 will report the change in parameters for each iteration, a value equal to 1 will report when each iteration is finished, a value equal to 0 will return no output.}

\item{tol}{Convergence criteria. Default is 1e-3.}

\item{min_iter}{Minimum number of iterations before returning approximation. Default is 5.}

\item{max_iter}{Maximum number of iterations before returning approximation. Default is 100.}

\item{scale}{Controls whether the data should be scaled prior to performing CAVI.}
}
\value{
A list  containing the the relevant parameters of the mean-field approximation.  The components of the list are:
\item{\code{mean_lambda}}{The mean of the factor loadings. A matrix of dimension \eqn{P \times J}.}
\item{\code{var_lambda}}{The variance of thefactor loadings. A list of length \eqn{P}, each entry is a matrix of dimension \eqn{K \times j} corresponding to the covariance of the common factor loadings of row \eqn{p}.}
\item{\code{shape_psi}}{The shape parameter of the gamma distribution about the inverse idiosyncratic errors. A vector of length \eqn{P}.}
\item{\code{rate_psi}}{The rate parameter of the gamma distribution about the inverse idiosyncratic errors. A vector of length \eqn{P}.}
\item{\code{shape_psi}}{The mean of the inverse of Psi, corresponding to idiosyncratic error. A vector of length \eqn{P}.}
\item{\code{mean_l}}{The mean of the study-specific latent factors associated with the common factor loadings. A list of length \eqn{S}, each entry is a list of length \eqn{N_s} where each entry is a vector of length \eqn{J}.}
}
\description{
The function implements the CAVI algorithm described in Hansen et al. (2023).
The code will return the relevant parameters of the mean-field
approximation described in the main paper.
}
\references{
Hansen, B., Avalos-Pacheco, A., Massimiliano, R., De Vito, R. (2023).
Fast Variational Inference for Bayesian Factor Analysis in Single and Multi-Study Settings.
Submitted manuscript.
}
