% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi_msfa.R
\name{svi_msfa}
\alias{svi_msfa}
\title{Stochastic Variational Inference for Bayesian Multi-Study Factor Analysis}
\usage{
svi_msfa(
  X_s,
  K,
  J_s,
  batch_prop = 0.2,
  fr = 0.75,
  delay = 1,
  verbose = 1,
  tol = 0.005,
  min_iter = 5,
  max_iter = 100,
  scale = TRUE
)
}
\arguments{
\item{X_s}{List of length \eqn{S}{S}, corresponding to number of different studies to be analyzed.
Each element of the list contains a data matrix, with the same number of columns \eqn{P}{P} for all the studies.}

\item{K}{Number of common factors.}

\item{J_s}{Number of study-specific factors. A vector of positive integers of length \eqn{S}{S}.}

\item{batch_prop}{Proportion of data to be sampled during each iteration of SVI. Default is 0.2.}

\item{fr}{Forgetting rate used in the SVI step size schedule. Default is 0.51. See Hoffman et al. (2013) for more details.}

\item{delay}{Delay parameter used in the SVI step size schedule. Default is 10. See Hoffman et al. (2013) for more details.}

\item{verbose}{Desired verbosity of output printed out each iteration. A value greater than 1 will report the change in parameters for each iteration, a value equal to 1 will report when each iteration is finished, a value equal to 0 will return no output.}

\item{tol}{Convergence criteria. Default is 1e-3.}

\item{min_iter}{Minimum number of iterations before returning approximation. Default is 5.}

\item{max_iter}{Maximum number of iterations before returning approximation. Default is 100.}

\item{scale}{Controls whether the data should be scaled prior to performing CAVI.}
}
\value{
A list  containing the the relevant parameters of the mean-field approximation.  The components of the list are:
\item{\code{mean_phi}}{The mean of the Common factor loadings. A matrix of dimension \eqn{P \times K}.}
\item{\code{var_phi}}{The variance of the Common factor loadings. A list of length \eqn{P}, each entry is a matrix of dimension \eqn{K \times K} corresponding to the covariance of the common factor loadings of row \eqn{p}.}
\item{\code{mean_lambda}}{The mean of the study specific factor loadings. A list of length \eqn{S}, each entry is a matrix of dimension \eqn{P \times J_s}.}
\item{\code{var_lambda}}{The variance of the Common factor loadings. A list of length \eqn{S}, each entry is a list of length \eqn{P}, which contains a matrix of dimension \eqn{J_s \times J_s} corresponding to the covariance of the study specific factor loadings of row \eqn{p}.}
\item{\code{shape_psi}}{The shape parameter of the gamma distribution about the inverse idiosyncratic errors. A list of length \eqn{S}, each entry is a vector of length \eqn{P}}
\item{\code{rate_psi}}{The rate parameter of the gamma distribution about the inverse idiosyncratic errors. A list of length \eqn{S}, each entry is a vector of length \eqn{P}}
\item{\code{mean_f}}{The mean of the study-specific latent factors associated with the common factor loadings. A list of length \eqn{S}, each entry is a list of length \eqn{N_s} where each entry is a vector of length \eqn{K}.}
\item{\code{mean_l}}{The mean of the study-specific latent factors associated with the study specific factor loadings. A list of length \eqn{S}, each entry is a list of length \eqn{N_s} where each entry is a vector of length \eqn{J_s}.}
}
\description{
The function implements the SVI algorithm described in Hansen et al. (2023).
The code will return the relevant parameters of the mean-field
approximation described in the main paper.
}
\references{
Hansen, B., Avalos-Pacheco, A., Massimiliano, R., De Vito, R. (2023).
Fast Variational Inference for Bayesian Factor Analysis in Single and Multi-Study Settings.
Submitted manuscript.

Hoffman, M. D. (2013). Stochastic Variational Inference. Journal of Machine Learning Research, 14, 1303–1347.
}
